Move node_context.h to public, format
