+++ include/ccf/user_frontend.h
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "ccf/common_endpoint_registry.h"
-#include "enclave/node_context.h"
+#include "ccf/node_context.h"
 #include "node/rpc/frontend.h"
 #include "service/network_tables.h"
