+++ src/endpoints/common_endpoint_registry.cpp
@@ -9,7 +9,7 @@
 #include "ccf/http_consts.h"
 #include "ccf/http_query.h"
 #include "ccf/json_handler.h"
-#include "enclave/node_context.h"
+#include "ccf/node_context.h"
 #include "node/rpc/call_types.h"
 #include "node/rpc/serialization.h"
 #include "node/tx_receipt.h"
